feat: 合并房间消息
